refactoring out assistance view
